+++ sphinx/source/changelog.rst
@@ -9,6 +9,8 @@ Changelog
 
 3.3.2beta
 ---------
+* Added the ``-p`` option to :ref:`skool2html.py` (to show the path to the
+  skoolkit package directory)
 * Fixed the appearance of transparent table cells in HTML output 
 * Added `jet_set_willy-dark.css` (to complete the 'dark' theme for that
   disassembly)